requirements.txt
# Core MCP
mcp>=0.1.0
fastmcp>=2.0.0

# ML Models (Local)
transformers>=4.30.0
torch>=2.0.0
tokenizers>=0.13.0
scikit-learn>=1.3.0

# HTTP Client
aiohttp>=3.8.0
requests>=2.28.0

# Data Processing
pandas>=1.5.0
numpy>=1.21.0

# Configuration
pydantic>=1.10.0
python-dotenv>=0.19.0

# Utilities
json5>=0.9.0

# Date/Time
python-dateutil>=2.8.0

# URL Processing
urllib3>=1.26.0

# Logging
loguru>=0.6.0

# Testing
pytest>=7.0.0
pytest-asyncio>=0.21.0 

# Core async file operations
aiofiles>=23.0.0

# HTTP requests and web scraping
requests>=2.28.0
beautifulsoup4>=4.11.0

# RSS feed parsing
feedparser>=6.0.0

# Data processing
pandas>=1.5.0
numpy>=1.24.0

# Date utilities
python-dateutil>=2.8.0

# Async support
asyncio>=3.4.0

loguru
aiofiles